+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@
 ## Fixed
 - Fixed various memory leaks.
 - Drop-in no longer overrides the log level in case of debug builds.
+- Address Lookup not displaying validation error on Card Component when no address has been selected.
 
 ## Changed
 - Flags are replaced by ISO codes in the phone number inputs (affected payment methods: MB Way, Pay Easy, Convenience Stores Japan, Online Banking Japan and Seven-Eleven).